pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "evenet-tools"
version = "0.1.0"
description = "Tools for EveNet: Download pretrained models and launch fine-tuning"
requires-python = ">=3.12"
dependencies = [
    "evenet>=0.1.2",
    "huggingface-hub>=0.19.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "requests>=2.31.0"
]

[project.scripts]
eve-finetune = "evenet_tools.cli:main"
